+++ tests/test_bandpass.py
@@ -54,7 +54,7 @@ def test_Bandpass_basic():
                         blue_limit=400, red_limit=550),
         galsim.Bandpass(galsim.LookupTable([3000,8700], [0.3, 0.87], interpolant='linear'),
                           wave_type='Angstroms').truncate(400,550),
-        galsim.Bandpass(galsim.LookupTable([100, 400-1.e-10, 400, 550, 550+1.e-10, 900],
+        galsim.Bandpass(galsim.LookupTable([100, 400-1.e-12, 400, 550, 550+1.e-12, 900],
                                            [0., 0., 0.4, 0.55, 0., 0.], interpolant='linear')),
     ]
     k1 = len(b_list)
@@ -68,15 +68,16 @@ def test_Bandpass_basic():
         b_list[4].thin(),
         b_list[5].thin(),
         b_list[6].thin(),
-        b_list[6].thin(preserve_range=True),
+        b_list[6].thin(preserve_range=False),
         b_list[7].thin(),
-        b_list[11].thin(),
-        b_list[11].thin(preserve_range=True),
+        b_list[11].thin(preserve_range=False, trim_zeros=False),
+        b_list[11].thin(preserve_range=False),
+        b_list[11].thin()
     ]
 
     for k,b in enumerate(b_list):
         print k,' b = ',b
-        if k not in [k1-1, len(b_list)-1]:
+        if k not in [k1-1, len(b_list)-2, len(b_list)-1]:
             np.testing.assert_almost_equal(b.blue_limit, 400, decimal=12)
             np.testing.assert_almost_equal(b.red_limit, 550, decimal=12)
         np.testing.assert_almost_equal(b(400), 0.4, decimal=12)
@@ -276,9 +277,43 @@ def test_ne():
     t2 = time.time()
     print 'time for %s = %.2f' % (funcname(), t2-t1)
 
+
+def test_thin():
+    import time
+    t1 = time.time()
+
+    s = galsim.SED('1', flux_type='fphotons')
+    bp = galsim.Bandpass(os.path.join(datapath, 'LSST_r.dat'))
+    flux = s.calculateFlux(bp)
+    print "Original number of bandpass samples = ",len(bp.wave_list)
+    for err in [1.e-2, 1.e-3, 1.e-4, 1.e-5]:
+        print "Test err = ",err
+        thin_bp = bp.thin(rel_err=err, preserve_range=True, fast_search=False)
+        thin_flux = s.calculateFlux(thin_bp)
+        thin_err = (flux-thin_flux)/flux
+        print "num samples with preserve_range = True, fast_search = False: ",len(thin_bp.wave_list)
+        print "realized error = ",(flux-thin_flux)/flux
+        thin_bp = bp.thin(rel_err=err, preserve_range=True)
+        thin_flux = s.calculateFlux(thin_bp)
+        thin_err = (flux-thin_flux)/flux
+        print "num samples with preserve_range = True: ",len(thin_bp.wave_list)
+        print "realized error = ",(flux-thin_flux)/flux
+        assert np.abs(thin_err) < err, "Thinned bandpass failed accuracy goal, preserving range."
+        thin_bp = bp.thin(rel_err=err, preserve_range=False)
+        thin_flux = s.calculateFlux(thin_bp)
+        thin_err = (flux-thin_flux)/flux
+        print "num samples with preserve_range = False: ",len(thin_bp.wave_list)
+        print "realized error = ",(flux-thin_flux)/flux
+        assert np.abs(thin_err) < err, "Thinned bandpass failed accuracy goal, w/ range shrinkage."
+
+    t2 = time.time()
+    print 'time for %s = %.2f' % (funcname(), t2-t1)
+
+
 if __name__ == "__main__":
     test_Bandpass_basic()
     test_Bandpass_mul()
     test_Bandpass_div()
     test_Bandpass_wave_type()
     test_ne()
+    test_thin()