+++ tests/test_interpolatedimage.py
@@ -1166,11 +1166,20 @@ def test_ne():
     perturb_image.array[64, 64] *= 1000
     obj2 = galsim.InterpolatedImage(perturb_image, calculate_maxk=False, calculate_stepk=False)
 
+    # These tests won't always work if astropy < 1.0.6 has been imported, so look for that.
+    import sys
+    if 'astropy' in sys.modules:
+        import astropy  # Just b/c someone imported it, doesn't mean we can see it yet.
+        from distutils.version import LooseVersion
+        if LooseVersion(astropy.__version__) < LooseVersion('1.0.6'):
+            return
+
     with galsim.utilities.printoptions(threshold=128*128):
-        assert repr(obj1) != repr(obj2)
+        assert repr(obj1) != repr(obj2), "Reprs unexpectedly agree: %r"%obj1
 
     with galsim.utilities.printoptions(threshold=1000):
-        assert repr(obj1) == repr(obj2)
+        assert repr(obj1) == repr(obj2), "Reprs disagree: repr(obj1)=%r\nrepr(obj2)=%r"%(
+                obj1, obj2)
 
     assert obj1 != obj2